feat: Overhaul setup and bring it into the 2020s
For a long while we've been meaning to setup the latest version of
powerline. We also wanted to manage our tmux related plug-in in a more
manageable away. Getting away from using git submodules and using
something that is easier to interact with.

We're now using `tpm` to handle this. To install our pacjages we can now
call `<C-a>i` which will leverage `tpm` and install all of our plug-ins.

We also take the time to update our themes and use "catppuccin". This is
now working with `erikw/tmux-powerline`.

The final thing we've added is the ability to use `fzf` from within
`tmux`.

Although this commit makes quite a few changes, it is now inline with
`neovim` and our kitty setup. Allowing us to have a seamless experiences
between our power tools.

Signed-off-by: Yomi Colledge <baphled@boodah.net>